// Models/people.ts
export interface PeopleModel {
  name: string
  height: number | string
  mass: number | string
  hair_color: string
  skin_color: string
  eye_color: string
  birth_year: string
  gender: string
  homeworld: string
  films: string[]
  starships: string[]
}

// "name": "Luke Skywalker",
// 	"height": "172",
// 	"mass": "77",
// 	"hair_color": "blond",
// 	"skin_color": "fair",
// 	"eye_color": "blue",
// 	"birth_year": "19BBY",
// 	"gender": "male",
// 	"homeworld": "https://swapi.dev/api/planets/1/",
// 	"films": [
// 		"https://swapi.dev/api/films/1/",
// 		"https://swapi.dev/api/films/2/",
// 		"https://swapi.dev/api/films/3/",
// 		"https://swapi.dev/api/films/6/"
// 	],
// 	"species": [],
// 	"vehicles": [
// 		"https://swapi.dev/api/vehicles/14/",
// 		"https://swapi.dev/api/vehicles/30/"
// 	],
// 	"starships": [
// 		"https://swapi.dev/api/starships/12/",
// 		"https://swapi.dev/api/starships/22/"
// 	],
// 	"created": "2014-12-09T13:50:51.644000Z",
// 	"edited": "2014-12-20T21:17:56.891000Z",
// 	"url": "https://swapi.dev/api/people/1/"
// }

//{
//   "name": "Darth Vader", 
//   "height": "202", 
//   "mass": "136", 
//   "hair_color": "none", 
//   "skin_color": "white", 
//   "eye_color": "yellow", 
//   "birth_year": "41.9BBY", 
//   "gender": "male", 
//   "homeworld": "https://swapi.dev/api/planets/1/", 
//   "films": [
//       "https://swapi.dev/api/films/1/", 
//       "https://swapi.dev/api/films/2/", 
//       "https://swapi.dev/api/films/3/", 
//       "https://swapi.dev/api/films/6/"
//   ], 
//   "species": [], 
//   "vehicles": [], 
//   "starships": [
//       "https://swapi.dev/api/starships/13/"
//   ], 
//   "created": "2014-12-10T15:18:20.704000Z", 
//   "edited": "2014-12-20T21:17:50.313000Z", 
//   "url": "https://swapi.dev/api/people/4/"
// }, 

// {
//   "name": "Obi-Wan Kenobi", 
//   "height": "182", 
//   "mass": "77", 
//   "hair_color": "auburn, white", 
//   "skin_color": "fair", 
//   "eye_color": "blue-gray", 
//   "birth_year": "57BBY", 
//   "gender": "male", 
//   "homeworld": "https://swapi.dev/api/planets/20/", 
//   "films": [
//       "https://swapi.dev/api/films/1/", 
//       "https://swapi.dev/api/films/2/", 
//       "https://swapi.dev/api/films/3/", 
//       "https://swapi.dev/api/films/4/", 
//       "https://swapi.dev/api/films/5/", 
//       "https://swapi.dev/api/films/6/"
//   ], 
//   "species": [], 
//   "vehicles": [
//       "https://swapi.dev/api/vehicles/38/"
//   ], 
//   "starships": [
//       "https://swapi.dev/api/starships/48/", 
//       "https://swapi.dev/api/starships/59/", 
//       "https://swapi.dev/api/starships/64/", 
//       "https://swapi.dev/api/starships/65/", 
//       "https://swapi.dev/api/starships/74/"
//   ], 
//   "created": "2014-12-10T16:16:29.192000Z", 
//   "edited": "2014-12-20T21:17:50.325000Z", 
//   "url": "https://swapi.dev/api/people/10/"
// }

// {
//   "name": "Chewbacca", 
//   "height": "228", 
//   "mass": "112", 
//   "hair_color": "brown", 
//   "skin_color": "unknown", 
//   "eye_color": "blue", 
//   "birth_year": "200BBY", 
//   "gender": "male", 
//   "homeworld": "https://swapi.dev/api/planets/14/", 
//   "films": [
//       "https://swapi.dev/api/films/1/", 
//       "https://swapi.dev/api/films/2/", 
//       "https://swapi.dev/api/films/3/", 
//       "https://swapi.dev/api/films/6/"
//   ], 
//   "species": [
//       "https://swapi.dev/api/species/3/"
//   ], 
//   "vehicles": [
//       "https://swapi.dev/api/vehicles/19/"
//   ], 
//   "starships": [
//       "https://swapi.dev/api/starships/10/", 
//       "https://swapi.dev/api/starships/22/"
//   ], 
//   "created": "2014-12-10T16:42:45.066000Z", 
//   "edited": "2014-12-20T21:17:50.332000Z", 
//   "url": "https://swapi.dev/api/people/13/"
// }